sync_adshao_go_binance (#5)

+++ v2/order_service.go
@@ -186,7 +186,7 @@ type CreateOrderResponse struct {
 
 // Fill may be returned in an array of fills in a CreateOrderResponse.
 type Fill struct {
-	TradeID         int    `json:"tradeId"`
+	TradeID         int64  `json:"tradeId"`
 	Price           string `json:"price"`
 	Quantity        string `json:"qty"`
 	Commission      string `json:"commission"`
